+++ CHANGELOG.md
@@ -46,6 +46,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * Added packages mailx, ksh and nmon
 * Added [SECURITY.md](SECURITY.md)
 * Added script `start_sshd.sh`
+* Added support for `-l|--lts` as kernel in [upload_iso.sh](upload_iso.sh)
 
 ### Changed in unreleased
 
@@ -53,6 +54,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   * Default path for old ZBM is now `/mnt/efi...`
   * Default path for new ZBM is now basepath of the called script
 * Moved from build image `archlinux:latest` to `archlinux/archlinux:latest` to fix issues like [archinstal#2443](https://github.com/archlinux/archinstall/issues/2443)
+* Moved archzfs mirrorlist to dedicated file `/etc/pacman.d/archzfs`
 
 ## [2.11.1](https://github.com/stevleibelt/arch-linux-live-cd-iso-with-zfs/tree/2.11.1) - 20231111